Try to add sqlite3 database
